Add PowerShellScript
Add preview on Providers page and copy to clipboard funcionality, they are both the same text.

The code is replicated now cause I couldn't figure out using a variable for the script, there is probably a little work to be made in the renderer to make it fully functional.

Only one difference between the code for the Callout and the code for the copytoclipboard: For the callout the line breaks are ´</br>´ and for the copytoclipboard the linebreaks are `\n`