IDP-5164 , IDP-5174
To Fix IDP-5174
Change validation for dataStores field and groups inputs field.
  * On dataStores, add "and is.notEmpty" to the "validate" field
  * On groups, add "and is.notEmpty" after the "is.required" on "validate" field
    Note: Some templates where using another type of "validate" and this didn't pressent the bug,  for example AppDynamics.json

To Fix IDP-5164:
Change the constrain rule for the "validation" section of the template, on "requireSigningCert > "nested" > "signingCert" > "constrain" change "value" for "serialNo".

Note that this fix the blocker part of the bug, probably the value not showing needs to be handle from outside the template